# specs/005-devexpress-reports/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DevExpress Reports API
  description: |
    API contracts for the DevExpress Reports integration in Herbal Medicine ERP.

    This specification covers two API surfaces:
    1. **Next.js API** - Report management, permissions, and data endpoints
    2. **ASP.NET Core API** - DevExpress reporting engine endpoints

    Authentication uses JWT tokens passed in the Authorization header.
  version: 1.0.0
  contact:
    name: Herbal Medicine ERP
servers:
  - url: http://localhost:3000/api
    description: Next.js API (Development)
  - url: http://localhost:5000
    description: ASP.NET Core Reporting Backend (Development)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from ERP authentication
  schemas:
...
    ReportPermission:
      type: object
      properties:
        id:
          type: integer
        templateId:
          type: integer
        role:
          type: string
          enum: [admin, manager, production, qc, warehouse, purchasing, sales, user]
        canView:
          type: boolean
          default: true
        canDesign:
          type: boolean
          default: false
        canExport:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
      required:
        - templateId
        - role
    ReportPermissionUpdate:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [admin, manager, production, qc, warehouse, purchasing, sales, user]
              canView:
                type: boolean
              canDesign:
                type: boolean
              canExport:
                type: boolean
            required:
              - role
      required:
        - permissions

    # --- Report Execution (Audit) ---
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
        message:
          type: string
security:
  - bearerAuth: []
paths:
...
  /reports/templates/{code}/publish:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string

    post:
      tags:
        - Report Templates
      summary: Publish a template
      description: Makes template available to users with view permission
      responses:
        '200':
          description: Template published
        '403':
          description: No design permission
        '404':
          description: Template not found
...
  /reports/templates/{code}/permissions:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Report Permissions
      summary: Get template permissions
      responses:
        '200':
          description: Permission list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ReportPermission'
        '403':
          description: No admin permission
        '404':
          description: Template not found

    put:
      tags:
        - Report Permissions
      summary: Update template permissions
      description: Replaces all permissions for the template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportPermissionUpdate'
      responses:
        '200':
          description: Permissions updated
        '403':
          description: No admin permission
        '404':
          description: Template not found

  # ===========================================
  # NEXT.JS API - REPORT DATA ENDPOINTS
  # ===========================================
  /reports/data/inventory-valuation:
    get:
      tags:
        - Report Data
      summary: Inventory valuation data
      description: Data source for inventory valuation reports
      parameters:
        - name: warehouseId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: asOfDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Inventory data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    itemCode:
                      type: string
                    itemName:
                      type: string
                    category:
                      type: string
                    warehouse:
                      type: string
                    quantity:
                      type: number
                    unitCost:
                      type: number
                    totalValue:
                      type: number
                    lotCount:
                      type: integer
...